openfl-tutorials/interactive_api/Pytorch_Market_Re-ID/envoy_folder/requirements.txt
gdown==3.13.0
Pillow==8.3.1